+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/AbstractSpotlessMojo.java
@@ -36,11 +36,13 @@ import org.eclipse.aether.repository.RemoteRepository;
 import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.Provisioner;
+import com.diffplug.spotless.maven.cpp.Cpp;
 import com.diffplug.spotless.maven.generic.Format;
 import com.diffplug.spotless.maven.generic.LicenseHeader;
 import com.diffplug.spotless.maven.java.Java;
 import com.diffplug.spotless.maven.kotlin.Kotlin;
 import com.diffplug.spotless.maven.scala.Scala;
+import com.diffplug.spotless.maven.xml.Xml;
 
 public abstract class AbstractSpotlessMojo extends AbstractMojo {
 
@@ -85,6 +87,12 @@ public abstract class AbstractSpotlessMojo extends AbstractMojo {
 
 	@Parameter
 	private Kotlin kotlin;
+
+	@Parameter
+	private Xml xml;
+
+	@Parameter
+	private Cpp cpp;
 
 	protected abstract void process(List<File> files, Formatter formatter) throws MojoExecutionException;
 
@@ -144,7 +152,7 @@ public abstract class AbstractSpotlessMojo extends AbstractMojo {
 	}
 
 	private List<FormatterFactory> getFormatterFactories() {
-		return Stream.concat(formats.stream(), Stream.of(java, scala, kotlin))
+		return Stream.concat(formats.stream(), Stream.of(java, scala, kotlin, cpp, xml))
 				.filter(Objects::nonNull)
 				.collect(toList());
 	}