+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/MavenProvisioner.java
@@ -15,10 +15,6 @@
  */
 package com.diffplug.spotless.maven;
 
-import static java.util.stream.Collectors.toSet;
-
-import java.util.Objects;
-
 import com.diffplug.spotless.Provisioner;
 
 /** Maven integration for Provisioner. */
@@ -26,10 +22,6 @@ public class MavenProvisioner {
 	private MavenProvisioner() {}
 
 	public static Provisioner create(ArtifactResolver artifactResolver) {
-		Objects.requireNonNull(artifactResolver);
-
-		return mavenCoords -> mavenCoords.stream()
-				.flatMap(coord -> artifactResolver.resolve(coord).stream())
-				.collect(toSet());
+		return artifactResolver::resolve;
 	}
 }